Migrate stale float-index tests to integer; expand bool advanced indices
The branch enforces integer indices in take/advanced-indexing (matching
NumPy and PyTorch, both of which reject float indices). Several legacy
tests created indices via mx.nd.array / mx.np.arange, which default to
float32, and relied on the old auto-cast. Pin those indices to int64 to
reflect the test's clear integer intent:
- test_ndarray_take, test_ndarray_fluent
- test_sparse_take
- test_grad_graph_change
- test_numpy_interoperability linalg.norm take workloads

Also expand boolean masks in _advanced_index_to_array to integer
positions via nonzero (NumPy semantics) instead of rejecting them, so
mixed boolean indexing (e.g. a[None, mask]) works again
(test_boolean_index_list).

+++ tests/python/unittest/test_gluon.py
@@ -1838,7 +1838,7 @@ def test_grad_graph_change():
             row = array.take(index)
             return row, index
     array = mx.np.arange(3)
-    index = mx.np.array([2])
+    index = mx.np.array([2], dtype='int64')
     array.attach_grad()
     model = Model()
     model.hybridize(inline_limit=0)
